@@ -15,30 +15,36 @@
     "watch-js": "nodemon dist/app.js"
   },
   "devDependencies": {
-    "concurrently": "^5.3.0",
-    "nodemon": "^2.0.4"
-  },
-  "dependencies": {
     "@koa/cors": "^3.1.0",
-    "@typegoose/typegoose": "^7.3.2",
+    "@typegoose/typegoose": "^7.4.1",
     "@types/dotenv": "^8.2.0",
     "@types/jsonwebtoken": "^8.5.0",
-    "@types/koa": "^2.11.3",
-    "@types/lodash": "^4.14.159",
+    "@types/koa": "^2.11.6",
+    "@types/lodash": "^4.14.162",
     "@types/mongoose": "^5.7.36",
-    "@types/node": "^14.6.0",
-    "axios": "^0.19.2",
+    "@types/node": "^14.14.3",
+    "concurrently": "^5.3.0",
+    "nodemon": "^2.0.6"
+  },
+  "dependencies": {
+    "@types/koa-bodyparser": "^4.3.0",
+    "axios": "^0.21.0",
     "dotenv": "^8.2.0",
     "facebook-node-sdk": "^0.2.0",
-    "joi": "^17.2.0",
+    "joi": "^17.3.0",
     "js-yaml": "^3.14.0",
     "jsonwebtoken": "^8.5.1",
     "koa": "^2.13.0",
-    "koa-bodyparser-ts": "^0.3.1",
+    "koa-bodyparser": "^4.3.0",
     "koa-router-ts": "^3.0.0",
     "lodash": "^4.17.20",
-    "mongoose": "^5.10.0",
-    "node-telegram-login": "^1.0.0",
-    "typescript": "^3.9.7"
+    "module-alias": "^2.2.2",
+    "mongoose": "^5.10.10",
+    "typescript": "^4.0.3"
+  },
+  "_moduleAliases": {
+    "@controllers": "dist/controllers",
+    "@helpers": "dist/helpers",
+    "@models": "dist/models"
   }
 }
